+++ CHANGELOG.md
@@ -14,6 +14,9 @@
   - on parser reset next status report will print WCO
   - modified dual endstop behavior when dual endstops option is not active
 
+### Fixed
+  - clearing interpolator now also resets dss previous value and clears running segment pointer to prevent step contamination from canceled motions (#124)
+
 ## [1.3.4] - 2022-01-10
 µCNC version 1.3.4 adds a few improvements and also fixes some issues with inverse feedrate mode `G93` and realtime feed overrides.